AVFoundation Test
I did some test how to accelerte the slow capturing of CGWindowListCreateImage by using AVCaptureScreenInput.
Even it is now more performant in theory this method did not work. Unfortunly AVCaptureScreenInput only grab the screen content after compositing and the output window of Dalton Lens can not be excluded like in the CGWindowListCreateImage method. This leeds to a bad unwanted visual feedback loop. I just create this branch to save the new code parts I tried even if this was a dead end.