experiments for Protanomaly
Some experiments I did to simulate different levels of Protanomaly within applyProtanomalyRgb() and try a different approach in daltonizeV1() to compensate the Protanomaly by not ignoring the red channel in compensation. This makes dark reds brighter and also make green birther as well. With this approach the daltonize correction will not longer make all reds just "look pink" and it is still possible to better differentiate the "protan confusion lines" for a protanomaly person.